requirements.txt
flask
requests
slackclient
slackeventsapi